// src/interfaces/ICalculate.ts
interface ICalculate {
    id?: number;
    date: string;
    salary: number;
    transfers: number;
    outcome1: number;
    outcome2: number;
    outcome3: number;
    outcome4: number;
    outcome5: number;
    outcome6: number;
    outcome7: number;
    outcome8: number;
    outcome9: number;
    outcome10: number;
    outcome11: number;
    outcome12: number;
}

export default ICalculate;